Create README.md
Mr BluJobs Is basically an android platform purely coded in Java that offers a variety of Tech Jobs all Over The world.Intent behind it was to show a couple of elements and revolve around what has to be offered in the field fetching this data from Github jobs.
Disclaimer GitHub Jobs is deprecated! New jobs will not be posted from May 19, 2021. It will shut down entirely on August 19, 2021.
Read more in the GitHub blog post.
All in all its basically a platform to showcase what you can do with Java on your way out to fully migrate to Kotlin